rename main page component
