+++ app/jenkins-for-jira-ui/src/components/ServerManagement/ServerManagement.tsx
@@ -3,11 +3,11 @@ import PageHeader from '@atlaskit/page-header';
 import { ButtonGroup } from '@atlaskit/button';
 import Button from '@atlaskit/button/standard-button';
 import { headerContainer } from '../JenkinsServerList/JenkinsServerList.styles';
-import { mainPageContainer } from './MainPage.styles';
+import { mainPageContainer } from './ServerManagement.styles';
 import { ConnectionPanel } from '../ConnectionPanel/ConnectionPanel';
 import { TopPanel } from './TopPanel/TopPanel';
 
-const MainPage = (): JSX.Element => {
+const ServerManagement = (): JSX.Element => {
 	const pageHeaderActions = (
 		<ButtonGroup>
 			{/* TODO - add onClick event */}
@@ -33,4 +33,4 @@ const MainPage = (): JSX.Element => {
 	);
 };
 
-export { MainPage };
+export { ServerManagement };
