+++ app/jenkins-for-jira-ui/src/components/ServerManagement/TopPanel/TopPanel.tsx
@@ -7,7 +7,7 @@ import {
 	topPanelContentHeaderContainer,
 	topPanelImgContainer,
 	topPanelImg
-} from '../MainPage.styles';
+} from '../ServerManagement.styles';
 import { PluginIcon } from '../../icons/PluginIcon';
 
 const TopPanel = (): JSX.Element => {
